+++ sample.env
@@ -6,3 +6,4 @@ CONFIG_BACKEND_URL="https://api.rudderstack.com"
 REDIS_HOST = localhost
 REDIS_PORT = 6379
 REDIS_PASSWORD = 123
+useRedisDatabase = true
