+++ src/util/redisConnector.js
@@ -1,7 +1,6 @@
 const redis = require('redis');
 const log = require('../logger');
 
-let a = 1;
 const host = process.env.REDIS_HOST || 'localhost';
 const port = process.env.REDIS_PORT || 6379;
 const password = process.env.REDIS_PASSWORD || '';
@@ -15,13 +14,12 @@ const DBConnector = {
   }),
   async getRedisInstance() {
     DBConnector.redisInstance.on('connect', () => {
-      log.info(`Redis Connected! id: ${a}`);
+      log.info(`Redis Connected!`);
     });
     if (!DBConnector.redisInstance?.isOpen) {
       await DBConnector.redisInstance
         .connect()
         .then((res) => {
-          a += 1;
         })
         .catch((err) => {
           log.info(`err happened while connecting to redis ${err}`);
