+++ test/__mocks__/data/sources/shopify/response.json
@@ -1,5 +1,8 @@
 {
     "shopify_test1": {
         "anonymousId": "anon_shopify_test1"
+    },
+    "shopify_test2": {
+        "anonymousId": "anon_shopify_test2"
     }
 }
