+++ test/__tests__/redis.test.js
@@ -1,25 +1,28 @@
+const integration = "redis";
+const name = "Redis";
+
 const fs = require("fs");
 const path = require("path");
 const version = "v0";
 
-const sourcesList= ['Shopify']
-describe(`Source Tests`, () => {
-sourcesList.forEach((source)=>{
-  const testDataFile = fs.readFileSync(
-    path.resolve(__dirname, `./data/redis/${source}_source.json`)
-  );
-  const data = JSON.parse(testDataFile);
-  const transformer = require(`../../src/${version}/sources/${source}/transform`);
-  
-  data.forEach((dataPoint, index) => {
-    it(`${index}. ${source} - ${dataPoint.description}`, async() => {
-      try {
-        const output = await transformer.process(dataPoint.input);
-        expect(output).toEqual(dataPoint.output);
-      } catch (error) {
-        expect(error.message).toEqual(dataPoint.output.error);
-      }
-    });
+const transformer = require(`../../src/${version}/destinations/${integration}/transform`);
+
+const inputDataFile = fs.readFileSync(
+  path.resolve(__dirname, `./data/${integration}_input.json`)
+);
+const outputDataFile = fs.readFileSync(
+  path.resolve(__dirname, `./data/${integration}_output.json`)
+);
+const inputData = JSON.parse(inputDataFile);
+const expectedData = JSON.parse(outputDataFile);
+
+inputData.forEach((input, index) => {
+  it(`${name} Tests: payload - ${index}`, () => {
+    try {
+      const output = transformer.process(input);
+      expect(output).toEqual(expectedData[index]);
+    } catch (error) {
+      expect(error.message).toEqual(expectedData[index].error);
+    }
   });
-})
-})
+});
